(WIP) Some more refactor to prepare to expand hunks down
