--- app/src/ui/diff/diff-expansion.ts
@@ -1,11 +1,0 @@
-import { DiffHunk, DiffLine } from '../../models/diff'
-
-export function getDiffTextFromHunks(hunks: ReadonlyArray<DiffHunk>) {
-  // Grab all hunk lines and rebuild the diff text from it
-  const newDiffLines = hunks.reduce<ReadonlyArray<DiffLine>>(
-    (result, hunk) => result.concat(hunk.lines),
-    []
-  )
-
-  return newDiffLines.map(diffLine => diffLine.text).join('\n')
-}
