+++ app/src/ui/diff/text-diff.tsx
@@ -25,7 +25,6 @@ import {
   findInteractiveDiffRange,
   lineNumberForDiffLine,
   DiffRangeType,
-  diffHunkForIndex,
   diffLineInfoForIndex,
 } from './diff-explorer'
 
@@ -45,12 +44,15 @@ import { showContextualMenu } from '../main-process-proxy'
 import { IMenuItem } from '../../lib/menu-item'
 import { enableDiscardLines } from '../../lib/feature-flag'
 import { canSelect } from './diff-helpers'
-import { getDiffTextFromHunks } from './diff-expansion'
+import {
+  expandTextDiffHunk,
+  ExpansionKind,
+  getDiffTextFromHunks,
+  getHunkExpansionInfo,
+} from './text-diff-expansion'
 
 /** The longest line for which we'd try to calculate a line diff. */
 const MaxIntraLineDiffStringLength = 4096
-
-const DiffExpansionDistance = 20
 
 // This is a custom version of the no-newline octicon that's exactly as
 // tall as it needs to be (8px) which helps with aligning it on the line.
@@ -90,10 +92,8 @@ interface ISelection {
   readonly isSelected: boolean
 }
 
-type ExpansionKind = 'up' | 'down'
-
 interface IExpansion {
-  readonly lineNumber: number
+  readonly hunk: DiffHunk
   readonly kind: ExpansionKind
 }
 
@@ -572,132 +572,18 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
       return
     }
 
-    const hunk = diffHunkForIndex(diff.hunks, this.expansion.lineNumber)
-
-    if (hunk === null) {
-      return
-    }
-
-    const hunkIndex = diff.hunks.indexOf(hunk)
-    if (hunkIndex === -1) {
-      return
-    }
-
-    // Grab the hunk line of the hunk to expand
-    const diffHunkLine = hunk.lines[0]
-    if (!diffHunkLine || diffHunkLine.type !== DiffLineType.Hunk) {
-      return
-    }
-
-    const newLineNumber = hunk.header.newStartLine
-    const oldLineNumber = hunk.header.oldStartLine
-
-    const isExpandingUp = this.expansion.kind === 'up'
-    const [from, to] = isExpandingUp
-      ? [newLineNumber - DiffExpansionDistance - 1, newLineNumber - 1]
-      : [newLineNumber + 1, newLineNumber + DiffExpansionDistance + 1]
-
-    const newLines = this.newContentLines.slice(
-      Math.max(from, 0),
-      Math.min(to, this.newContentLines.length)
+    const updatedDiff = expandTextDiffHunk(
+      diff,
+      this.expansion.hunk,
+      this.expansion.kind,
+      this.newContentLines
     )
-    const numberOfLinesToAdd = newLines.length
-
-    // Nothing to do here
-    if (numberOfLinesToAdd === 0) {
-      return
-    }
-
-    // Create the DiffLine instances using the right line numbers.
-    const newLineDiffs = newLines.map((line, index) => {
-      const newNewLineNumber = isExpandingUp
-        ? newLineNumber - (numberOfLinesToAdd - index)
-        : newLineNumber + 1 + index
-      const newOldLineNumber = isExpandingUp
-        ? oldLineNumber - (numberOfLinesToAdd - index)
-        : oldLineNumber + 1 + index
-
-      // We need to prepend a space before the line text to match the diff
-      // output.
-      return new DiffLine(
-        ' ' + line,
-        DiffLineType.Context,
-        newNewLineNumber,
-        newOldLineNumber,
-        false
-      )
-    })
-
-    // Update the resulting hunk header with the new line count
-    const newHunkHeader = new DiffHunkHeader(
-      hunk.header.oldStartLine - numberOfLinesToAdd,
-      hunk.header.oldLineCount + numberOfLinesToAdd,
-      hunk.header.newStartLine - numberOfLinesToAdd,
-      hunk.header.newLineCount + numberOfLinesToAdd
-    )
-
-    // Create a new Hunk header line, except if we're expanding up and we
-    // reached the top of the file. Store in an array to make it easier to add
-    // later to the new list of lines.
-    // TODO: handle similar scenario when expanding down
-    const newDiffHunkLine =
-      isExpandingUp && from <= 0
-        ? []
-        : [
-            new DiffLine(
-              `@@ ${newHunkHeader.toDiffRepresentation()} @@`,
-              DiffLineType.Hunk,
-              diffHunkLine.oldLineNumber,
-              diffHunkLine.newLineNumber,
-              diffHunkLine.noTrailingNewLine
-            ),
-          ]
-
-    const allButDiffLine = hunk.lines.slice(1)
-
-    // Update the diff lines of the hunk with the new lines
-    const updatedHunkLines = isExpandingUp
-      ? [...newDiffHunkLine, ...newLineDiffs, ...allButDiffLine]
-      : [...newDiffHunkLine, ...allButDiffLine, ...newLineDiffs]
-
-    const numberOfNewDiffLines = updatedHunkLines.length - hunk.lines.length
-
-    // Update the hunk with all the new info (header, lines, start/end...)
-    const updatedHunk = new DiffHunk(
-      newHunkHeader,
-      updatedHunkLines,
-      hunk.unifiedDiffStart,
-      hunk.unifiedDiffEnd + numberOfNewDiffLines
-    )
-
-    const previousHunks = diff.hunks.slice(0, hunkIndex)
-
-    // Grab the hunks after the current one, and update their start/end
-    const followingHunks = diff.hunks
-      .slice(hunkIndex + 1)
-      .map(
-        hunk =>
-          new DiffHunk(
-            hunk.header,
-            hunk.lines,
-            hunk.unifiedDiffStart + numberOfNewDiffLines,
-            hunk.unifiedDiffEnd + numberOfNewDiffLines
-          )
-      )
-
-    // TODO: merge hunks
-
-    // Create the new list of hunks of the diff, and the new diff text
-    const newHunks = [...previousHunks, updatedHunk, ...followingHunks]
-    const newDiffText = getDiffTextFromHunks(newHunks)
-
-    const updatedDiff = {
-      ...diff,
-      text: newDiffText,
-      hunks: newHunks,
-    }
 
     this.expansion = null
+
+    if (updatedDiff === undefined) {
+      return
+    }
 
     this.setState({
       diff: updatedDiff,
@@ -1093,32 +979,10 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
     const isIncludeable = diffLine.isIncludeableLine()
     const isIncluded = isIncludeable && this.isIncluded(index)
     const hover = isIncludeable && inSelection(this.hunkHighlightRange, index)
-    let isExpandableDown = false
-    let isExpandableUp = false
-    let isExpandableBoth = false
-    let isExpandableShort = false
-
-    if (diffLine.type === DiffLineType.Hunk) {
-      const hunkIndex = hunks.indexOf(hunk)
-      const previousHunk = hunks[hunkIndex - 1]
-      const distanceToPrevious =
-        previousHunk === undefined
-          ? Infinity
-          : hunk.header.oldStartLine -
-            previousHunk.header.oldStartLine -
-            previousHunk.header.oldLineCount
-
-      if (hunkIndex === 0) {
-        isExpandableUp = true
-      } else if (distanceToPrevious <= DiffExpansionDistance) {
-        isExpandableShort = true
-      } else if (hunkIndex === hunks.length - 1 && hunk.lines.length === 1) {
-        // TODO: we need to add a dummy hunk at the end to
-        isExpandableDown = true
-      } else {
-        isExpandableBoth = true
-      }
-    }
+    const hunkExpansionInfo =
+      diffLine.type === DiffLineType.Hunk
+        ? getHunkExpansionInfo(hunks, hunk)
+        : undefined
 
     return {
       'diff-line-gutter': true,
@@ -1129,10 +993,10 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
       'read-only': this.props.readOnly,
       'diff-line-selected': isIncluded,
       'diff-line-hover': hover,
-      'expandable-down': isExpandableDown,
-      'expandable-up': isExpandableUp,
-      'expandable-both': isExpandableBoth,
-      'expandable-short': isExpandableShort,
+      'expandable-down': hunkExpansionInfo?.isExpandableDown === true,
+      'expandable-up': hunkExpansionInfo?.isExpandableUp === true,
+      'expandable-both': hunkExpansionInfo?.isExpandableBoth === true,
+      'expandable-short': hunkExpansionInfo?.isExpandableShort === true,
       includeable: isIncludeable && !this.props.readOnly,
     }
   }
@@ -1148,7 +1012,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
 
     marker.addEventListener(
       'mousedown',
-      this.onDiffLineGutterMouseDown.bind(this, index, diffLine)
+      this.onDiffLineGutterMouseDown.bind(this, index, hunk, diffLine)
     )
 
     const oldLineNumber = document.createElement('div')
@@ -1170,7 +1034,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
     hunkExpandUpHandle.classList.add('hunk-expand-up-handle')
     hunkExpandUpHandle.addEventListener(
       'mousedown',
-      this.onHunkExpandHandleMouseDown.bind(this, index, 'up')
+      this.onHunkExpandHandleMouseDown.bind(this, hunk, 'up')
     )
     marker.appendChild(hunkExpandUpHandle)
 
@@ -1178,7 +1042,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
     hunkExpandDownHandle.classList.add('hunk-expand-down-handle')
     hunkExpandDownHandle.addEventListener(
       'mousedown',
-      this.onHunkExpandHandleMouseDown.bind(this, index, 'down')
+      this.onHunkExpandHandleMouseDown.bind(this, hunk, 'down')
     )
     marker.appendChild(hunkExpandDownHandle)
 
@@ -1188,7 +1052,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
   }
 
   private onHunkExpandHandleMouseDown = (
-    index: number,
+    hunk: DiffHunk,
     kind: ExpansionKind,
     ev: MouseEvent
   ) => {
@@ -1204,7 +1068,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
 
     ev.preventDefault()
 
-    this.startExpansion(index, kind)
+    this.startExpansion(hunk, kind)
   }
 
   private updateGutterMarker(
@@ -1287,6 +1151,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
 
   private onDiffLineGutterMouseDown = (
     index: number,
+    hunk: DiffHunk,
     diffLine: DiffLine,
     ev: MouseEvent
   ) => {
@@ -1308,7 +1173,7 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
     if (diffLine.type === DiffLineType.Hunk) {
       ev.preventDefault()
       // TODO: detect expand up or down (or both?)
-      this.startExpansion(index, 'up')
+      this.startExpansion(hunk, 'up')
       return
     }
 
@@ -1321,13 +1186,13 @@ export class TextDiff extends React.Component<ITextDiffProps, ITextDiffState> {
     this.startSelection(file, diff.hunks, index, 'range')
   }
 
-  private startExpansion(lineNumber: number, kind: ExpansionKind) {
+  private startExpansion(hunk: DiffHunk, kind: ExpansionKind) {
     if (this.expansion !== null) {
       this.cancelExpansion()
     }
 
     this.expansion = {
-      lineNumber,
+      hunk,
       kind,
     }
